+++ mywords-flutter/lib/pages/dict_database.dart
@@ -12,6 +12,7 @@ import 'package:mywords/util/get_scaffold.dart';
 import 'package:mywords/util/path.dart';
 import 'package:mywords/util/util.dart';
 import '../libso/types.dart';
+import '../util/block_show.dart';
 
 class DictDatabase extends StatefulWidget {
   const DictDatabase({super.key});
@@ -20,41 +21,6 @@ class DictDatabase extends StatefulWidget {
   State createState() {
     return _State();
   }
-}
-
-/// blockShowDialog 阻塞 试验
-Future<void> blockShowDialog(BuildContext context, Future<void> future) {
-  return showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        const waiting = UnconstrainedBox(
-          child: SizedBox(
-            width: 50,
-            height: 50,
-            child: CircularProgressIndicator(),
-          ),
-        );
-        return FutureBuilder(
-            future: future,
-            builder: (BuildContext context, AsyncSnapshot<dynamic> snapshot) {
-              //snapshot就是_calculation在时间轴上执行过程的状态快照
-              switch (snapshot.connectionState) {
-                case ConnectionState.none:
-                  return waiting;
-                case ConnectionState.waiting:
-                  return waiting;
-                case ConnectionState.active:
-                  return waiting;
-                case ConnectionState.done:
-                  Future.delayed(const Duration(milliseconds: 0), () {
-                    if (context.mounted) {
-                      Navigator.of(context).pop();
-                    }
-                  });
-                  return const Text("");
-              }
-            });
-      });
 }
 
 class _State extends State<DictDatabase> {
